chore(core): move integration test to unit tests
TICKET: BG-38247
